Clarified readme, trimmed headings, made options into a list. Updated demo.css to be leaner. Updated demo files to include project gutenberg citation.
Updated plugin file to hand $(window) down to the placesaver object to be
stored in this.context, so that the helper functions don't have to call
$(window) directly. This allows a dev to bind the plugin to an element
other than $(window), should they ever actually have to.

Also, a bug was fixed in endReading() which caused a new saved place to
be created if the reader reached the clearElement and then scrolled back
up. It now also clears the currentScrollPosition value and unbinds the plugin
from the context.